requirements/local.txt
django==4.2.11
django-environ
djangorestframework
django-cors-headers
django-autoslug
django-countries
django-phonenumber-field
phonenumbers
drf-yasg
Pillow
django-admin-tailwind
argon2-cffi
python-dotenv
pytz
redis
celery
flower
djoser==2.2.2
django-taggit==5.0.1
django-filter==24.1
django-celery-beat==2.6.0
django-celery-email==3.0.0
cloudinary==1.39.1
whitenoise==5.3.0

psycopg2-binary

flake8==6.0.0
black==23.1.0
isort==5.12.0
watchfiles==0.18.1

pytest-django==4.5.2
pytest-factoryboy==2.5.1
Faker==18.3.1
pytest-cov==4.0.0
django-admin-interface